+++ tests/testthat/test-plotCohortTiming.R
@@ -46,11 +46,17 @@ test_that("plotCohortTiming, boxplot", {
   timing1 <- summariseCohortTiming(cdm$table,
     restrictToFirstEntry = TRUE
   )
-  boxplot1 <- plotCohortTiming(timing1,
+  expect_warning(boxplot1 <- plotCohortTiming(timing1,
     facet = "cdm_name",
     colour = c("cohort_name_reference", "cohort_name_comparator"),
     uniqueCombinations = TRUE
-  )
+  ))
+
+  expect_warning(plotCohortTiming(timing1,
+                                   facet = "cdm_name",
+                                   colour = c("cohort_name_reference", "cohort_name_comparator"),
+                                   uniqueCombinations = TRUE
+  ))
   # expect_true(all(c("q0", "q25", "q50", "q75", "q100") %in% colnames(boxplot1$data)))
   # expect_true(all(c("Cohort 1", "Cohort 2") %in% boxplot1$data$cohort_name_reference))
   # expect_true(all(c("Cohort 2", "Cohort 3", "Cohort 4") %in% boxplot1$data$cohort_name_comparator))
@@ -58,10 +64,10 @@ test_that("plotCohortTiming, boxplot", {
   expect_true(all(c("gg", "ggplot") %in% class(boxplot1)))
   # expect_true(boxplot1$labels$fill == "group")
 
-  boxplot2 <- plotCohortTiming(timing1,
+  expect_warning(boxplot2 <- plotCohortTiming(timing1,
     colour = c("cohort_name_comparator"),
     uniqueCombinations = FALSE
-  )
+  ))
   # expect_true(all(c("Cohort 1", "Cohort 2") %in% boxplot2$data$cohort_name_reference))
   # expect_true(all(c("Cohort 1", "Cohort 2", "Cohort 3", "Cohort 4") %in% boxplot2$data$cohort_name_comparator))
   expect_true(all(c("gg", "ggplot") %in% class(boxplot2)))
@@ -76,11 +82,11 @@ test_that("plotCohortTiming, boxplot", {
     strata = list("age_group", c("age_group", "sex")),
     restrictToFirstEntry = FALSE
   )
-  boxplot3 <- plotCohortTiming(timing3,
+  expect_warning(boxplot3 <- plotCohortTiming(timing3,
     colour = c("age_group", "sex"),
     facet = c("age_group", "sex"),
     uniqueCombinations = FALSE
-  )
+  ))
   # expect_true(all(c("Cohort 1", "Cohort 2") %in% boxplot3$data$cohort_name_reference))
   # expect_true(all(c("Cohort 1", "Cohort 2", "Cohort 3", "Cohort 4") %in% boxplot3$data$cohort_name_comparator))
   expect_true(all(c("gg", "ggplot") %in% class(boxplot3)))
@@ -136,34 +142,34 @@ test_that("plotCohortTiming, density", {
   )
 
   timing1 <- summariseCohortTiming(cdm$table, restrictToFirstEntry = FALSE)
-  density1 <- plotCohortTiming(timing1,
+  expect_warning(density1 <- plotCohortTiming(timing1,
     plotType = "densityplot",
     facet = NULL,
     colour = c("cohort_name_reference", "cohort_name_comparator"),
     uniqueCombinations = TRUE
-  )
+  ))
 
   # expect_true(all(c("plot_id", "timing_label", "color_var", "x", "y", ".group") %in% colnames(density1$data)))
   expect_true(all(c("gg", "ggplot") %in% class(density1)))
   # expect_true(density1$labels$fill == "color_var")
 
-  density2 <- plotCohortTiming(timing1,
+  expect_warning(density2 <- plotCohortTiming(timing1,
     plotType = "densityplot",
     colour = c("cohort_name_comparator"),
     facet = c("cdm_name", "cohort_name_reference"),
     uniqueCombinations = FALSE
-  )
+  ))
   # expect_true(all(c("plot_id", "timing_label", "x", "y", ".group") %in% colnames(density2$data)))
   expect_true(all(c("gg", "ggplot") %in% class(density2)))
   # expect_null(density2$labels$fill)
 
   timing2 <- summariseCohortTiming(cdm$table, estimates = "density")
-  density4 <- plotCohortTiming(timing2,
+  expect_warning(density4 <- plotCohortTiming(timing2,
     plotType = "densityplot",
     facet = NULL,
     colour = c("cohort_name_reference", "cohort_name_comparator"),
     uniqueCombinations = TRUE
-  )
+  ))
   expect_true(all(c("gg", "ggplot") %in% class(density4)))
   # expect_true(all(is.na(density4$data$q50)))
 
@@ -178,12 +184,12 @@ test_that("plotCohortTiming, density", {
     restrictToFirstEntry = FALSE
   )
 
-  density3 <- plotCohortTiming(timing3,
+  expect_warning(density3 <- plotCohortTiming(timing3,
     plotType = "densityplot",
     colour = c("age_group", "sex"),
     facet = c("cohort_name_reference", "cohort_name_comparator"),
     uniqueCombinations = FALSE
-  )
+  ))
   # expect_true(all(c("plot_id", "timing_label", "color_var", "x", "y", ".group") %in% colnames(density3$data)))
   expect_true(all(c("gg", "ggplot") %in% class(density3)))
   # expect_true(all(unique(density3$data$color_combined) %in% c("Overall", "0 to 40", "0 to 40 and female",
@@ -228,22 +234,21 @@ test_that("plotCohortTiming, density x axis", {
 
   # first
   timing <- summariseCohortTiming(cdm$cohort)
-  expect_no_error(p <- plotCohortTiming(timing, plotType = "densityplot"))
+  expect_warning(p <- plotCohortTiming(timing, plotType = "densityplot"))
   xLimits <- ggplot2::ggplot_build(p)$layout$panel_params[[1]]$x.range
   expect_true(xLimits[2] - xLimits[1] >= 5)
-  expect_no_error(p <- plotCohortTiming(timing, plotType = "densityplot", timeScale = "years"))
+  expect_warning(p <- plotCohortTiming(timing, plotType = "densityplot", timeScale = "years"))
   xLimits <- ggplot2::ggplot_build(p)$layout$panel_params[[1]]$x.range
   expect_true(xLimits[2] - xLimits[1] >= 5/365)
 
   # all
   timing <- summariseCohortTiming(cdm$cohort, restrictToFirstEntry = FALSE)
-  expect_no_error(p <- plotCohortTiming(timing, plotType = "densityplot"))
+  expect_warning(p <- plotCohortTiming(timing, plotType = "densityplot"))
   xLimits <- ggplot2::ggplot_build(p)$layout$panel_params[[1]]$x.range
   expect_true(xLimits[2] - xLimits[1] >= 5)
-  expect_no_error(p <- plotCohortTiming(timing, plotType = "densityplot", timeScale = "years"))
+  expect_warning(p <- plotCohortTiming(timing, plotType = "densityplot", timeScale = "years"))
   xLimits <- ggplot2::ggplot_build(p)$layout$panel_params[[1]]$x.range
   expect_true(xLimits[2] - xLimits[1] >= 5/365)
 
   PatientProfiles::mockDisconnect(cdm = cdm)
 })
-